v0.3.0
"Remember Me" option is now followed when logging in
Linking functionality is now detached from the login page.
Redirecting to last known place now works when logging in or linking Steam account.
Fixed registering not checking for existing user inbetween the initial authentication and the new user being inserted.
Added some more error handling to linking (if the user is logged out or the Steam account they're trying to link is already linked).